+++ tests/Bowl/BowlWithUnionTest.php
@@ -5,7 +5,7 @@
  *
  * LICENSE
  *
- * This source file is subject to the MIT license and the version 3 of the GPL3
+ * This source file is subject to the MIT license
  * license that are bundled with this package in the folder licences
  * If you did not receive a copy of the license and are unable to
  * obtain it through the world-wide-web, please send an email
@@ -79,49 +79,5 @@ EOF;
             $this->getMapping(),
             'bowlClass'
         );
-    }
-
-    public function testExceptionOnExecuteWithBadChef()
-    {
-        if (PHP_VERSION_ID < 80000) {
-            self::markTestSkipped('Only for PHP 8');
-
-            return;
-        }
-
-        parent::testExceptionOnExecuteWithBadChef();
-    }
-
-    public function testExceptionOnExecuteWithBadWorkPlan()
-    {
-        if (PHP_VERSION_ID < 80000) {
-            self::markTestSkipped('Only for PHP 8');
-
-            return;
-        }
-
-        parent::testExceptionOnExecuteWithBadWorkPlan();
-    }
-
-    public function testExecute()
-    {
-        if (PHP_VERSION_ID < 80000) {
-            self::markTestSkipped('Only for PHP 8');
-
-            return;
-        }
-
-        parent::testExecute();
-    }
-
-    public function testExceptionWhenExecuteAndMissingAndIngredientInWorkPlan()
-    {
-        if (PHP_VERSION_ID < 80000) {
-            self::markTestSkipped('Only for PHP 8');
-
-            return;
-        }
-
-        parent::testExceptionWhenExecuteAndMissingAndIngredientInWorkPlan();
     }
 }